% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persons.R
\name{getPersonsByTask}
\alias{getPersonsByTask}
\title{Get responses from a syllabus}
\usage{
getPersonsByTask(task_ids, connStr)
}
\arguments{
\item{task_ids}{A vector of task ids.}

\item{connStr}{A connection string.}
}
\value{
A data frame with responses.
}
\description{
Get responses from a syllabus
}
\examples{
getPersonsByTask(taskids, 'mongodb://')
}
