% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syllabus.R
\name{getSyllabusByName}
\alias{getSyllabusByName}
\title{Get responses from a syllabus}
\usage{
getSyllabusByName(name, connStr)
}
\arguments{
\item{name}{The syllabus Name}

\item{connStr}{A connection string.}
}
\value{
A data frame with syllabus details
}
\description{
Get responses from a syllabus
}
\examples{
getSyllabus('ISW 2022-2023 Year 7 Sept', 'mongodb://')
}
