% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{getSyllabusResponses}
\alias{getSyllabusResponses}
\title{Get responses from a syllabus}
\usage{
getSyllabusResponses(syllabusId, connStr)
}
\arguments{
\item{syllabusId}{The syllabus ID.}

\item{connStr}{A connection string.}
}
\value{
A data frame with responses.
}
\description{
Get responses from a syllabus
}
\examples{
getSyllabusResponses('67059790-569b-440f-a858-bddabf313e07', 'mongodb://')
}
