% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{getTasksBySyllabus}
\alias{getTasksBySyllabus}
\title{Get tasks from a syllabus}
\usage{
getTasksBySyllabus(syllabusId, connStr)
}
\arguments{
\item{syllabusId}{The syllabus ID.}

\item{connStr}{A connection string.}
}
\value{
A data frame with task details
}
\description{
Get tasks from a syllabus
}
\examples{
getTasksBySyllabus('67059790-569b-440f-a858-bddabf313e07', 'mongodb://')
}
